+++ app.js
@@ -26,7 +26,7 @@ _.extend(global.settings, env);
  
 // kick off controller
 if ( ! global.settings.base_url ) global.settings.base_url = '/api/*';
-var app = require(global.settings.app_root + '/app/controllers/app');
+var app = require(global.settings.app_root + '/app/controllers/app')();
 app.listen(global.settings.node_port, global.settings.node_host, function() {
   console.log("CartoDB SQL API listening on " +
       global.settings.node_host + ":" + global.settings.node_port +