+++ test/acceptance/app.auth.test.js
@@ -1,7 +1,7 @@
 require('../helper');
 require('../support/assert');
 
-var app    = require(global.settings.app_root + '/app/controllers/app')
+var app    = require(global.settings.app_root + '/app/controllers/app')()
     , assert = require('assert')
     , tests  = module.exports = {}
     , querystring = require('querystring');