+++ test/acceptance/app.test.js
@@ -16,7 +16,7 @@ require('../helper');
 require('../support/assert');
 
 
-var app    = require(global.settings.app_root + '/app/controllers/app')
+var app    = require(global.settings.app_root + '/app/controllers/app')()
     , assert = require('assert')
     , querystring = require('querystring')
     , _ = require('underscore')
@@ -26,8 +26,6 @@ var app    = require(global.settings.app_root + '/app/controllers/app')
     , Step = require('step')
     ;
 
-// allow lots of emitters to be set to silence warning
-app.setMaxListeners(0);
 
 suite('app.test', function() {