+++ test/acceptance/x-cache-channel.js
@@ -1,7 +1,7 @@
 require('../helper');
 require('../support/assert');
 
-var app    = require(global.settings.app_root + '/app/controllers/app')
+var app    = require(global.settings.app_root + '/app/controllers/app')()
     , assert = require('assert')
     , querystring = require('querystring')
     , _ = require('underscore')
@@ -102,6 +102,5 @@ test('survives partial transactions', function(done) {
         done();
     });
 });
-
 
 });